+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 - Update `RevitTest` version to `1.9.0`. 
 ### Console
-- Update `AppName` to include `Revit` version allow run at the same time.
+- Update `AppName` to include process `Id` to allow run multiple process.
 ### Tests
 - Add `RevitTestFixture` to test `TestFixture` in `ricaun.NUnit` version `1.5.0`.